Move publishers to their own file
